Refactoring: Move mediator to mediator package.
