+++ core/src/main/java/bisq/core/proto/network/CoreNetworkProtoResolver.java
@@ -34,12 +34,12 @@ import bisq.core.dao.node.messages.GetBlocksRequest;
 import bisq.core.dao.node.messages.GetBlocksResponse;
 import bisq.core.dao.node.messages.NewBlockBroadcastMessage;
 import bisq.core.dispute.arbitration.Arbitrator;
-import bisq.core.dispute.arbitration.Mediator;
 import bisq.core.dispute.arbitration.messages.DisputeCommunicationMessage;
 import bisq.core.dispute.arbitration.messages.DisputeResultMessage;
 import bisq.core.dispute.arbitration.messages.OpenNewDisputeMessage;
 import bisq.core.dispute.arbitration.messages.PeerOpenedDisputeMessage;
 import bisq.core.dispute.arbitration.messages.PeerPublishedDisputePayoutTxMessage;
+import bisq.core.dispute.mediator.Mediator;
 import bisq.core.filter.Filter;
 import bisq.core.offer.OfferPayload;
 import bisq.core.offer.messages.OfferAvailabilityRequest;
