+++ core/src/main/java/bisq/core/trade/Trade.java
@@ -23,8 +23,8 @@ import bisq.core.btc.wallet.BtcWalletService;
 import bisq.core.btc.wallet.TradeWalletService;
 import bisq.core.dispute.arbitration.Arbitrator;
 import bisq.core.dispute.arbitration.ArbitratorManager;
-import bisq.core.dispute.arbitration.Mediator;
 import bisq.core.dispute.arbitration.messages.DisputeCommunicationMessage;
+import bisq.core.dispute.mediator.Mediator;
 import bisq.core.filter.FilterManager;
 import bisq.core.locale.CurrencyUtil;
 import bisq.core.monetary.Price;
