+++ core/src/test/java/bisq/core/arbitration/MediatorTest.java
@@ -17,7 +17,7 @@
 
 package bisq.core.arbitration;
 
-import bisq.core.dispute.arbitration.Mediator;
+import bisq.core.dispute.mediator.Mediator;
 
 import bisq.network.p2p.NodeAddress;
